# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions/{sessionId}/status": {
      "get": {
        "operationId": "StatusController_getStatuses",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status updates visible to the session, grouped by contact."
          }
        },
        "summary": "Get all contact status updates",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/{contactId}": {
      "get": {
        "operationId": "StatusController_getContactStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status updates from the requested contact."
          }
        },
        "summary": "Get status updates from a specific contact",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/{statusId}/media": {
      "get": {
        "operationId": "StatusController_getStatusMedia",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The status image/video bytes."
          },
          "404": {
            "description": "No stored media (text status, omitted, or expired)."
          }
        },
        "summary": "Stream a stored status media file",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/send-text": {
      "post": {
        "operationId": "StatusController_sendTextStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTextStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Text status posted. The recipients allow-list is honored on Baileys only; whatsapp-web.js broadcasts to the account's status-privacy audience."
          },
          "400": {
            "description": "Invalid request, or the post was blocked by a plugin."
          }
        },
        "summary": "Post a text status",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/send-image": {
      "post": {
        "operationId": "StatusController_sendImageStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendImageStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image status posted. The recipients allow-list is honored on Baileys only; whatsapp-web.js broadcasts to the account's status-privacy audience."
          },
          "400": {
            "description": "Neither url nor base64 provided, or the post was blocked by a plugin."
          },
          "413": {
            "description": "Base64 media exceeds MEDIA_DOWNLOAD_MAX_BYTES."
          }
        },
        "summary": "Post an image status",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/send-video": {
      "post": {
        "operationId": "StatusController_sendVideoStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendVideoStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Video status posted. The recipients allow-list is honored on Baileys only; whatsapp-web.js broadcasts to the account's status-privacy audience."
          },
          "400": {
            "description": "Neither url nor base64 provided, or the post was blocked by a plugin."
          },
          "413": {
            "description": "Base64 media exceeds MEDIA_DOWNLOAD_MAX_BYTES."
          }
        },
        "summary": "Post a video status",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/{statusId}": {
      "delete": {
        "operationId": "StatusController_deleteStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status deleted."
          }
        },
        "summary": "Delete own status",
        "tags": [
          "status"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.10.8",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "SendTextStatusDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Status text body.",
            "example": "Out for delivery 📦",
            "maxLength": 4096
          },
          "backgroundColor": {
            "type": "string",
            "description": "Background color (hex).",
            "example": "#25D366"
          },
          "font": {
            "type": "number",
            "description": "Font family index (0–5).",
            "example": 0,
            "minimum": 0,
            "maximum": 5
          },
          "recipients": {
            "description": "Recipient JIDs (0–256). WhatsApp Status is not posted to a group — use @c.us or @lid individuals. Required on the Baileys engine (it posts to exactly this allow-list); ignored by whatsapp-web.js, which broadcasts to the account's status-privacy audience — omit it there.",
            "example": [
              "628123456789@c.us"
            ],
            "maxItems": 256,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "text"
        ]
      },
      "StatusMediaInput": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Public http(s) URL of the media (server-fetched, SSRF-guarded).",
            "example": "https://example.com/banner.jpg"
          },
          "base64": {
            "type": "string",
            "description": "Base64-encoded media. Requires mimetype.",
            "example": "/9j/4AAQSkZJRg..."
          },
          "mimetype": {
            "type": "string",
            "description": "MIME type. Required when sending base64.",
            "example": "image/jpeg"
          }
        }
      },
      "SendImageStatusDto": {
        "type": "object",
        "properties": {
          "image": {
            "description": "Image source (URL or base64).",
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusMediaInput"
              }
            ]
          },
          "caption": {
            "type": "string",
            "description": "Optional caption.",
            "example": "New drop!",
            "maxLength": 1024
          },
          "recipients": {
            "description": "Recipient JIDs (0–256), @c.us or @lid. Required on the Baileys engine (it posts to exactly this allow-list); ignored by whatsapp-web.js, which broadcasts to the account's status-privacy audience — omit it there.",
            "example": [
              "628123456789@c.us"
            ],
            "maxItems": 256,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "image"
        ]
      },
      "SendVideoStatusDto": {
        "type": "object",
        "properties": {
          "video": {
            "description": "Video source (URL or base64).",
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusMediaInput"
              }
            ]
          },
          "caption": {
            "type": "string",
            "description": "Optional caption.",
            "example": "Demo",
            "maxLength": 1024
          },
          "recipients": {
            "description": "Recipient JIDs (0–256), @c.us or @lid. Required on the Baileys engine (it posts to exactly this allow-list); ignored by whatsapp-web.js, which broadcasts to the account's status-privacy audience — omit it there.",
            "example": [
              "628123456789@c.us"
            ],
            "maxItems": 256,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "video"
        ]
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}